refactor(payment-gated-subs): address PR review feedback
 ## Context

PR review identified four improvement areas: implicit coupling on
cancelation_reason, double rule evaluation, missing guard on pending
rule transition, and missing declined test coverage.

 ## Description

Extract ResolveSubscriptionStatusService from EvaluateService to
separate rule evaluation from subscription status resolution. This
eliminates the double evaluation when called from ResolveService.

Make cancelation_reason explicit in Payment::ResolveService by saving
it after ResolveSubscriptionStatusService runs rather than relying on
dirty attribute persistence through a downstream service.

Add guard in Payment::EvaluateService#transition_pending_rule that
raises ArgumentError when status is nil, preventing silent no-ops.

Add polymorphic evaluate! method to ActivationRule model for
type-specific dispatch without case statements.

Add test coverage for declined rule status.
