+++ app/services/subscriptions/activation_rules/evaluate_service.rb
@@ -18,7 +18,7 @@ module Subscriptions
           result.rules << rule
         end
 
-        resolve_subscription_status
+        ResolveSubscriptionStatusService.call!(subscription:)
 
         result.subscription = subscription
         result
@@ -27,25 +27,6 @@ module Subscriptions
       private
 
       attr_reader :subscription
-
-      def resolve_subscription_status
-        return unless subscription.incomplete?
-
-        if all_rules_satisfied?
-          Subscriptions::ActivateService.call!(subscription:)
-        elsif any_rule_failed?
-          subscription.mark_as_canceled!
-          SendWebhookJob.perform_after_commit("subscription.canceled", subscription)
-        end
-      end
-
-      def all_rules_satisfied?
-        subscription.activation_rules.all? { |rule| rule.satisfied? || rule.not_applicable? }
-      end
-
-      def any_rule_failed?
-        subscription.activation_rules.any? { |rule| rule.failed? || rule.expired? || rule.declined? }
-      end
     end
   end
 end
