+++ app/services/subscriptions/activation_rules/payment/evaluate_service.rb
@@ -38,6 +38,8 @@ module Subscriptions
         end
 
         def transition_pending_rule
+          raise ArgumentError, "status required to transition a pending rule" unless status.present?
+
           rule.public_send(:"#{status}!")
         end
