+++ app/services/subscriptions/activation_rules/resolve_subscription_status_service.rb
@@ -1,0 +1,40 @@
+# frozen_string_literal: true
+
+module Subscriptions
+  module ActivationRules
+    class ResolveSubscriptionStatusService < BaseService
+      Result = BaseResult[:subscription]
+
+      def initialize(subscription:)
+        @subscription = subscription
+        super
+      end
+
+      def call
+        return result.tap { result.subscription = subscription } unless subscription.incomplete?
+
+        if all_rules_satisfied?
+          Subscriptions::ActivateService.call!(subscription:)
+        elsif any_rule_failed?
+          subscription.mark_as_canceled!
+          SendWebhookJob.perform_after_commit("subscription.canceled", subscription)
+        end
+
+        result.subscription = subscription
+        result
+      end
+
+      private
+
+      attr_reader :subscription
+
+      def all_rules_satisfied?
+        subscription.activation_rules.all? { |rule| rule.satisfied? || rule.not_applicable? }
+      end
+
+      def any_rule_failed?
+        subscription.activation_rules.any? { |rule| rule.failed? || rule.expired? || rule.declined? }
+      end
+    end
+  end
+end
